Migrate CLI, Networking, and RepositoriesLive tests to hostless targets
Split the next three layers into SPM test targets (BrewCLITests,
BrewNetworkingTests, BrewRepositoriesLiveTests), re-pointing @testable imports
at the owning modules and the support libs. Relocate the cross-target URLProtocol
session stub and the removePersistedKeys UserDefaults helper into
BrewCoreTestSupport, and move cache/concurrency integration tests down to the
layer they exercise. 127 tests across 25 suites pass via `swift test` hostless.
