+++ Tests/BrewCLITests/PackageUninstallCommandTests.swift
@@ -1,47 +1,45 @@
 //
-//  PackageUpgradeCommandTests.swift
+//  PackageUninstallCommandTests.swift
 //  BrewTests
 //
 
-@testable import Brew
+@testable import BrewCLI
+import BrewCore
+import BrewCoreTestSupport
+import BrewServicesTestSupport
 import Foundation
 import Testing
 
-struct PackageUpgradeCommandTests {
-    @Test func `operation id from package matches package id`() {
-        let package = InstalledBrewPackage.fixture(name: "wget", kind: .formula)
-        #expect(BrewOperationID(package: package).packageID == .formula(name: "wget"))
-    }
-
-    @Test func `formula run invokes brew upgrade name`() async throws {
-        let runner = CapturingCommandRunner()
+struct PackageUninstallCommandTests {
+    @Test func `formula run invokes brew uninstall formula name`() async throws {
+        let runner = CapturingUninstallCommandRunner()
         let brewURL = URL(fileURLWithPath: "/opt/homebrew/bin/brew")
         let ctx = BrewCommandExecutionContext(
             commandRunner: runner,
             locator: BrewExecutableLocator(overrideURL: brewURL),
         )
         let package = InstalledBrewPackage.fixture(name: "git", kind: .formula)
-        try await PackageUpgradeCommand(package: package).run(in: ctx)
+        try await PackageUninstallCommand(package: package).run(in: ctx)
 
         #expect(await runner.lastExecutable == brewURL)
-        #expect(await runner.lastArguments == ["upgrade", "--formula", "git"])
+        #expect(await runner.lastArguments == ["uninstall", "--formula", "git"])
     }
 
-    @Test func `cask run invokes brew upgrade cask name`() async throws {
-        let runner = CapturingCommandRunner()
+    @Test func `cask run invokes brew uninstall cask name`() async throws {
+        let runner = CapturingUninstallCommandRunner()
         let brewURL = URL(fileURLWithPath: "/usr/local/bin/brew")
         let ctx = BrewCommandExecutionContext(
             commandRunner: runner,
             locator: BrewExecutableLocator(overrideURL: brewURL),
         )
         let package = InstalledBrewPackage.fixture(name: "Slack", kind: .cask)
-        try await PackageUpgradeCommand(package: package).run(in: ctx)
+        try await PackageUninstallCommand(package: package).run(in: ctx)
 
-        #expect(await runner.lastArguments == ["upgrade", "--cask", "Slack"])
+        #expect(await runner.lastArguments == ["uninstall", "--cask", "Slack"])
     }
 
     @Test func `nonzero exit throws BrewCommandError`() async {
-        let runner = CapturingCommandRunner()
+        let runner = CapturingUninstallCommandRunner()
         await runner.setOutput(
             CommandOutput(standardOutput: "", standardError: "blocked", terminationStatus: 1),
         )
@@ -51,14 +49,25 @@ struct PackageUpgradeCommandTests {
         )
         let package = InstalledBrewPackage.fixture(name: "x", kind: .formula)
         await #expect(throws: BrewCommandError.self) {
-            try await PackageUpgradeCommand(package: package).run(in: ctx)
+            try await PackageUninstallCommand(package: package).run(in: ctx)
         }
+    }
+
+    @Test func `kind and name initializer matches uninstall argv`() async throws {
+        let runner = CapturingUninstallCommandRunner()
+        let brewURL = URL(fileURLWithPath: "/opt/homebrew/bin/brew")
+        let ctx = BrewCommandExecutionContext(
+            commandRunner: runner,
+            locator: BrewExecutableLocator(overrideURL: brewURL),
+        )
+
+        try await PackageUninstallCommand(kind: .formula, name: "wget").run(in: ctx)
+        #expect(await runner.lastExecutable == brewURL)
+        #expect(await runner.lastArguments == ["uninstall", "--formula", "wget"])
     }
 }
 
-// MARK: - Test doubles
-
-private actor CapturingCommandRunner: BrewCommandRunning {
+private actor CapturingUninstallCommandRunner: BrewCommandRunning {
     private(set) var lastExecutable: URL?
     private(set) var lastArguments: [String]?
     private var output = CommandOutput(standardOutput: "", standardError: "", terminationStatus: 0)
