+++ Tests/BrewCLITests/SerialBrewCommandCenterTests.swift
@@ -3,7 +3,10 @@
 //  BrewTests
 //
 
-@testable import Brew
+@testable import BrewCLI
+import BrewCore
+import BrewCoreTestSupport
+import BrewServicesTestSupport
 import Foundation
 import Testing
