+++ Tests/BrewNetworkingTests/BrewAPIClientCatalogueURLSessionIntegrationTests.swift
@@ -3,7 +3,9 @@
 //  BrewTests
 //
 
-@testable import Brew
+import BrewCore
+import BrewCoreTestSupport
+@testable import BrewNetworking
 import Foundation
 import Testing
