+++ Tests/BrewRepositoriesLiveTests/BrewInstalledPackagesRepositoryTests.swift
@@ -3,10 +3,18 @@
 //  BrewTests
 //
 
-@testable import Brew
+import BrewCLI
+import BrewCore
+import BrewCoreTestSupport
+import BrewNetworking
+import BrewRepositories
+@testable import BrewRepositoriesLive
+import BrewRepositoriesTestSupport
+import BrewServicesTestSupport
 import Foundation
 import Testing
 
+// swiftlint:disable:next type_body_length
 struct BrewInstalledPackagesRepositoryTests {
     @Test @MainActor func `load returns sorted packages for mixed formula and cask json payload`() async throws {
         let json = """
